Add tracker (as python file and windows executable)
The get_active_app.py logs for each second: the timestamp and what app the user was on -this is then stored onto a JSON called data.json which stores what application the user is on every second, it then counts every iteration of each app (number of seconds) and stores that value onto iterations.json, the data.json is deleted every hour to keep the file from increasing, and is recreated automatically. iterations.json is deleted every 24 hours as it is no longer needed.